ci(release): bootstrap automated multi-arch release pipeline
 Set up end-to-end GitHub Actions automation so the first `v*` tag can
 build,
 package, and publish a full GitHub Release without manual steps.

 - add CI workflow for fmt/clippy/build checks on PRs and pushes
 - add release workflow for tagged/manual releases with:
   - mipsel + aarch64 builds
   - serialmux + windlass-bridge artifacts
   - SHA256 checksum generation and manifest
   - git-cliff-based changelog/release notes
   - GitHub Release asset publication
 - add nightly workflow with change detection and rolling nightly tag
   update
 - add cross-compilation config (`Cross.toml`, `.cargo/config.toml`) and
   Taskfile
   targets for Rust cross-builds and changelog generation
 - add CI/CD and release docs (`CI-CD*.md`, `RELEASE-QUICK-START.md`,
   `BUILDING.md`)
 - stop tracking generated binaries and ignore `build/*`
 - include minor Rust cleanups needed for stricter lint/build
   consistency

+++ serialmux-rs/src/protocol.rs
@@ -337,8 +337,13 @@ impl TxQueue {
 
         // SAFETY: `fd` is a valid open non-blocking file descriptor; the slice
         // `buf[head..head+avail]` lies within the allocated buffer.
-        let written =
-            unsafe { libc::write(fd, self.buf[head..].as_ptr() as *const libc::c_void, avail) };
+        let written = unsafe {
+            libc::write(
+                fd,
+                self.buf[head..].as_ptr() as *const libc::c_void,
+                avail as libc::size_t,
+            )
+        };
         if written < 0 {
             let e = std::io::Error::last_os_error();
             if e.raw_os_error()
